aider: Updated the README.md file for the `add_iac` role to reflect the role's new name `hashicorp`.
# Aider chat conversation:

USER: The readme needs to be updated as well.

ASSISTANT: To update the README.md file for the `add_iac` role to reflect the role's new name `hashicorp`, I will provide the diff for the changes. Please add the `vm-setup/roles/add_iac/README.md` file to the chat so I can make the necessary changes.

USER: that file is added

ASSISTANT: Here is the diff for the changes to update the `README.md` file for the `add_iac` role to reflect the role's new name `hashicorp`:

```diff
--- vm-setup/roles/add_iac/README.md
+++ vm-setup/roles/add_iac/README.md
@@ ... @@
 ## Role Name

-The `add_iac` role is responsible for setting up Infrastructure as Code (IaC)
+The `hashicorp` role is responsible for setting up Infrastructure as Code (IaC)
 tools on the target system. Currently, it focuses on adding the HashiCorp
 repository and installing HashiCorp products such as Terraform and Packer.

@@ ... @@
 ## Example Playbook

 Including an example of how to use your role (for instance, with variables
 passed in as parameters) is always nice for users too:

 ```yaml
 - hosts: servers
   roles:
-     - role: add_iac
+     - role: hashicorp
 ```
```

Please apply this diff to the `README.md` file located in the `vm-setup/roles/add_iac` directory. After applying the changes, you can proceed to rename the directory to `hashicorp` as you mentioned earlier.