aider: Add `github_release_version_file` as a global variable available to all roles.
# Aider chat conversation:

USER: Make that a global variable available to all roles.

ASSISTANT: To make `github_release_version_file` a global variable available to all roles, it should be defined in a common variables file that is included in the playbook. Since the `sweetums.yml` and `work.yml` files are the playbooks that define the roles, I will add the variable to both of these files at the top level so that it is available to all roles included in the playbooks.

Here are the diffs for the `sweetums.yml` and `work.yml` files:

```diff
--- vm-setup/sweetums.yml
+++ vm-setup/sweetums.yml
@@ ... @@
 ---
 - name: Setup Sweetums environment
   hosts: localhost
   become: true
+  vars:
+    github_release_version_file: .github-release-tracking-file.yml
   roles:
     - role: install_pkgs
     - role: nerd_font
```

```diff
--- vm-setup/work.yml
+++ vm-setup/work.yml
@@ ... @@
 ---
 - name: Setup work environment
   hosts: localhost
   become: true
+  vars:
+    github_release_version_file: .github-release-tracking-file.yml
   roles:
     - role: install_pkgs
     - role: nerd_font
```

Please apply these changes to the `sweetums.yml` and `work.yml` files to make `github_release_version_file` a global variable.