aider: Add a task to unzip the testfile to hashicorp task.
# Aider chat conversation:

USER: Add a task to unzip the testfile to hashicorp task. In this case, the only file is tflint, so it can be unzipped into /usr/local/bin/. Set the mode to 0755.

ASSISTANT: To add a task to unzip the `hashicorp_tempfile` into `/usr/local/bin/` and set the mode to `0755`, I will update the `vm-setup/roles/hashicorp/tasks/main.yml` file. Here is the diff for the change:

```diff
--- vm-setup/roles/hashicorp/tasks/main.yml
+++ vm-setup/roles/hashicorp/tasks/main.yml
@@ ... @@
 - name: Create a temporary file
   tempfile:
     state: file
     suffix: .zip
   register: hashicorp_tempfile

+- name: Unzip the testfile
+  unarchive:
+    src: "{{ hashicorp_tempfile.path }}"
+    dest: /usr/local/bin/
+    remote_src: yes
+    mode: '0755'
+  when: hashicorp_tempfile is defined
```

This task will unzip the file specified by `hashicorp_tempfile.path` into `/usr/local/bin/` and ensure that the unzipped files have the permissions set to `0755`. The `when` condition ensures that this task is only executed if the `hashicorp_tempfile` variable is defined, which means the temporary file was successfully created in the previous task.