aider: Create 'common_pkgs' role that calls 'install_pkgs' with default packages.
# Aider chat conversation:

USER: In the vm-setup directory, create a 'common_pkgs' role that calls the install_pkgs with the default packages from the install_pkgs role.

ASSISTANT: To create a 'common_pkgs' role that calls the 'install_pkgs' role with the default packages, I will create several new files within the `vm-setup/roles/common_pkgs` directory. These files will include:

1. `README.md` to describe the role.
2. `meta/main.yml` to define role dependencies.
3. `tasks/main.yml` to include the `install_pkgs` role and use its defaults.

Here are the diffs for these new files:

```diff
--- /dev/null
+++ vm-setup/roles/common_pkgs/README.md
@@ ... @@
+Role Name
+=========
+
+A brief description of the role goes here.
+
+Requirements
+------------
+
+Any pre-requisites that may not be covered by Ansible itself or the role should be mentioned here.
+
+Role Variables
+--------------
+
+None.
+
+Dependencies
+------------
+
+This role depends on the `install_pkgs` role.
+
+Example Playbook
+----------------
+
+Including an example of how to use your role (for instance, with variables passed in as parameters) is always nice for users too:
+
+    - hosts: servers
+      roles:
+         - { role: common_pkgs }
```

```diff
--- /dev/null
+++ vm-setup/roles/common_pkgs/meta/main.yml
@@ ... @@
+---
+galaxy_info:
+  author: Your Name
+  description: Common packages installation role
+  company: Your Company
+  license: license (MIT, GPL-2.0, etc)
+  min_ansible_version: 2.4
+  platforms:
+    - name: EL
+      versions:
+        - 7
+        - 8
+  galaxy_tags: []
+dependencies:
+  - role: install_pkgs
```

```diff
--- /dev/null
+++ vm-setup/roles/common_pkgs/tasks/main.yml
@@ ... @@
+---
+# This playbook includes the install_pkgs role to install default packages.
+
+- name: Install common packages using install_pkgs defaults
+  include_role:
+    name: install_pkgs
```

Please note that the `meta/main.yml` file assumes that the `install_pkgs` role is already present in the roles directory and that it has default variables defined that list the packages to be installed. If this is not the case, you will need to define those defaults or pass the package list explicitly.