aider: Corrected the structure of the `test-role.yml` file to define a playbook with the `docker` role for testing on `localhost`.
# Aider chat conversation:

USER: I ran this command:

 cd vm-setup && ansible-playbook -K test-role.yml

And got this output:

ERROR! A malformed role declaration was encountered.

The error appears to be in '/home/harleypig/projects/dotfiles/vm-setup/test-role.yml': line 2, column 3, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

---
- name: Test specific roles
  ^ here

ASSISTANT: It seems like there might be an issue with the structure of the `test-role.yml` file. To resolve the "malformed role declaration" error, the file should be properly formatted according to Ansible's expectations for a playbook. The playbook should start with a list of plays, each play should contain a `hosts` declaration, and roles should be listed under the `roles` key.

I will provide a diff to correct the structure of the `test-role.yml` file. If `test-role.yml` is not in the list of *read-write* files, please add it to the chat so I can make the necessary changes.

If `test-role.yml` is available for editing, here is the diff to correct its structure:

```diff
--- vm-setup/test-role.yml
+++ vm-setup/test-role.yml
@@ ... @@
 ---
-- name: Test specific roles
+- hosts: localhost
+  become: yes
+  roles:
+    - { role: docker }
```

This diff assumes that you want to test the `docker` role on the `localhost` and that you need to become a privileged user (usually root) to apply the changes. If you have a different role to test or different hosts, please provide that information.