aider: Added mkdir -p commands for XDG variables in shell-startup file.
# Aider chat conversation:

USER: Add mkdir -p commands for each of those xdg variables.

ASSISTANT: Here is the diff to add `mkdir -p` commands for each of the XDG variables in the `shell-startup` file:

```diff
--- shell-startup
+++ shell-startup
@@ ... @@
 # XDG Variables

 export XDG_CONFIG_HOME="$DOTFILES/config"
+mkdir -p "$XDG_CONFIG_HOME"
 export XDG_CACHE_HOME="$HOME/.cache"
+mkdir -p "$XDG_CACHE_HOME"
 export XDG_DATA_HOME="$DOTFILES/local_share"
+mkdir -p "$XDG_DATA_HOME"

 ##############################################################################
 # Functions
```

This diff will ensure that the directories for `XDG_CONFIG_HOME`, `XDG_CACHE_HOME`, and `XDG_DATA_HOME` are created if they do not already exist when the `shell-startup` script is run.