aider: Bind the `history-expand-line` command to `Alt+Space` in `inputrc`.
# Aider chat conversation:

USER: What setting in the inputrc does the same thing as Space:magic-space?

ASSISTANT: The `Space:magic-space` bind command in Bash is used to perform history expansion when the space key is pressed. This means that if you type a history expansion character (like `!`) followed by space, it will immediately expand to the relevant command from history.

In the `inputrc` file, there isn't a direct equivalent to `Space:magic-space` because `inputrc` does not support binding actions to space in the same way that Bash's `bind` command does. The `inputrc` file is used for configuring Readline, which is more focused on line editing and does not have a feature that directly corresponds to the `magic-space` functionality in Bash.

However, Readline does support history expansion, which is enabled by default. You can type a history expansion character and press Enter to perform the expansion. If you want to see the expanded command before executing it, you can set the `history-expand-line` Readline command to a key binding in `inputrc` and use that key to expand the history inline.

For example, you could bind the `history-expand-line` to the `Alt+Space` key combination in `inputrc` like this:

```diff
--- config/readline/inputrc
+++ config/readline/inputrc
@@ ... @@
+ "\e ": history-expand-line
```

This would allow you to press `Alt+Space` to expand the history, similar to how `Space:magic-space` works in Bash, but it's not exactly the same since it requires using the `Alt` key. There is no direct `inputrc` equivalent to the `bind` command's `magic-space`.