aider: Remove the `mark-directories` setting from the `inputrc` file.
# Aider chat conversation:

USER: For the completion related settings, are any of them setting default values?

ASSISTANT: To determine if any of the completion-related settings in the `config/readline/inputrc` file are set to their default values, we need to refer to the documentation or source code of the readline library. According to the readline documentation, here are the default values for the completion-related settings:

- `colored-completion-prefix`: This option is not mentioned in the readline manual, which means it is likely not a default setting provided by readline.
- `completion-ignore-case`: Off by default.
- `completion-map-case`: Off by default.
- `completion-prefix-display-length`: The default value is not specified in the readline manual.
- `mark-directories`: On by default.
- `mark-symlinked-directories`: Off by default.
- `show-all-if-ambiguous`: Off by default.
- `show-all-if-unmodified`: This option is not mentioned in the readline manual, which means it is likely not a default setting provided by readline.

Based on this information, the following settings are already set to their default values and can be removed from the `inputrc` file:

- `mark-directories`: On by default.

Here is the diff for removing this setting:

```diff
--- config/readline/inputrc
+++ config/readline/inputrc
@@ ... @@
-set mark-directories on
```

The other settings are either not default or it's unclear what their default values are, so they should be left as they are unless you have specific preferences or further information.